requirements.txt
absl-py==1.3.0
asttokens==2.1.0
backcall==0.2.0
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.6
cycler==0.11.0
decorator==5.1.1
executing==1.2.0
Flask==2.2.2
fonttools==4.38.0
google-auth==2.14.1
google-auth-oauthlib==0.4.6
grpcio==1.50.0
idna==3.4
importlib-metadata==5.0.0
ipython==8.6.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
numpy==1.23.4
oauthlib==3.2.2
opencv-python==4.6.0.66
packaging==21.3
pandas==1.5.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.3.0
prompt-toolkit==3.0.32
protobuf==3.20.3
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.9.3
seaborn==0.12.1
six==1.16.0
stack-data==0.6.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
thop==0.1.1.post2209072238
torch==1.13.0
torchvision==0.14.0
tqdm==4.64.1
traitlets==5.5.0
typing_extensions==4.4.0
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.2.2
zipp==3.10.0
